feat: add LLM to Zed editor
Still hate genAI, but have to get used to it cuz otherwise I won't have
a job apparently. Unfortunately the industry is moving towards it, and
we can only hope that there will soon come a time when we have ethically
trained LLMs. Until then, this is the sad reality.